pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "egile-agent-x-twitter"
version = "0.1.0"
description = "X/Twitter agent plugin for Egile Agent Core - integrates with egile-mcp-x-post-creator"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["agent", "plugin", "mcp", "twitter", "x", "social", "content"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "egile-agent-core @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-core",
    "egile-mcp-x-post-creator @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-mcp-x-post-creator",
    "agno>=2.3.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
all = [
    "egile-agent-core @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-agent-core",
    "egile-mcp-x-post-creator @ file:///C:/Users/jeanb/OneDrive/Documents/projects/egile-mcp-x-post-creator",
    "agno>=2.3.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[project.entry-points."egile_agent_core.plugins"]
xtwitter = "egile_agent_x_twitter.plugin:XTwitterPlugin"

[project.scripts]
xtwitter = "egile_agent_x_twitter.run_server:run_all"
xtwitter-mcp = "egile_agent_x_twitter.run_server:run_mcp_only"
xtwitter-agent = "egile_agent_x_twitter.run_server:run_agent_only"

[tool.hatch.build.targets.wheel]
packages = ["src/egile_agent_x_twitter"]

[tool.hatch.metadata]
allow-direct-references = true
